Basic structure of vision completed
Finished the structure of Align command. Still need to implement the ta, tx, and ty values into the targetPose to make the robot line up the AprilTag properly.

Fixed odometry issues in SwerveSubsytem and SwerveModule! (I think)

Created getPosition function in SwerveModule to have odometry work with the Align command. However, it doesn't work quite yet: I'm having some issues with how RelatvieEncoders work.

The variable "tv," aka if the limelight is targeting an apriltag, is now displayed on SmartDashboard.